feat(formatter): Enable adding links in terminal output
Parse syntax in config file format variables like:

  [[[link to something]($style)][https://example.com/${something}]]

which will produce terminal output using OSC (Operating System
Control) code 7 to indicate that there's a URI embedded in the output.
In the terminal output from starship the user will see:

  link to something

and the terminal will make that text into a clickable link to:

  https://example.com/<contents of something variable>

This could be useful, for instance, when adding a link to the
remote git repo for some of the various Starship git modules without
massively expanding the size of the prompt by displaying the text of
the link itself -- instead an appropriate symbol or shortened string
could be output and this would enable clicking or copying the link
in terminals such as those using the VTE library.

Note that this can also be (ab)used to make "tooltips" by ensuring
the "tooltip" does not conform to standard URL syntax.
(lacks a prefix like, for example, https://). For this use
imagine we have a glyph which has an arrow but the meaning of the
arrow may not be immediately obvious to the user. The module could
have in its format string::

  [[(${arrow_symbol})][a tooltip explaining symbol]]

which would show the tooltip if and only if the arrow symbol was
emitted and the user hovered their mouse over it.

Uses '[[' and ']]' to avoid altering the escaping requirements
which could break existing configs. This also somewhat resembles
Markdown syntax for "Reference-style Links" so it might feel more
natural to users familiar with Markdown.

For the purposes of parsing internals, the URL is like a TextGroup
except no styling is applied to the URL. Also, right now we allow
links within links because disallowing it would complicate the
formatting grammar and parsing. This should work just fine because
we only wrap TextSegments with URLSegments and leave existing
URLSegments in the parsed-then-flattened format alone. However
it's not optimal with respect to outputting ANSI/xterm escape codes
and it might be handy to employ the "id=foo" parameters to indicate
related links.