TODO fix(context): walk branches to find remote
Even in simple workflows local branches are often tracking the local
main/master branch. This means the results of
get_remote_repository_info() are often not useful or even unwanted.
While that may have motivated the "ignore_branches" configuration
option this presents an alternative: walk beyond the local branches
(usually just "master" or "main") until we reach a genuine remote
repository and branch.

It's also possible that In more complicated workflows users can have
multiple local branches the current branch is based on. So the
get_remote_*info() function can be something of a misnomer as it
will almost always return "." (local) as the name of the "remote",
especially on smaller and simpler projects.

So find the true remote by walking the fetch branches until the
remote's name is not ".". This leaves the ignore_branches configuration
intact since it's likely incorporated in many configurations at this
point.

TODO: Remove old code in comment
TODO: test this in combination with the ignore_branches configuration