feat(git_branch): Add $remote_symbol
Add a configuration table to map the FQDN portion of a URL to
a symbol, populate it with useful defaults, and make the result
of that lookup usable in the format string as a meta variable
($remote_symbol).

Keys starting with . are used when the hostname is the name
of the repository or project. For instance, the Linux kernel
repo on googlesource.com is found at kernel.goooglesource.com
so the key needed to match repos hosted there is .googlesource.com