Adapt wrap_*_for_shell to new escape sequences
Other CSI escape sequences start with \u{1b} yet are not terminated
by m (e.g. OSC sequences). This produces errors when those other
escape sequences are present in the strings we're building.

To make wrap_colorseq_for_shell() work with other sequences be more
specific by searching for the CSI SGR sequence string (ESC [) rather
than just the ESC character that it starts with.

As an added bonus, this simplifies the closure so we don't need
any external variables to track whether we're in an escape sequence
or not, uses fewer comparisons, and half as many format!() macro
applications.

A model in the Rust Playground which may help reviewers verify this
works as intended:
  https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=88db57134a559c12faa3a6bf1e61810c